Uplink console for both radio session and state session
