+++ ptest/radio_session.py
@@ -27,7 +27,7 @@ class RadioSession(object):
     '''
 
 
-    def __init__(self, device_name, imei, simulation_run_dir, tlm_config):
+    def __init__(self, device_name, imei, uplink_console, simulation_run_dir, tlm_config):
         '''
         Initializes state session with the Quake radio.
         '''
@@ -35,6 +35,9 @@ class RadioSession(object):
         # Device connection
         self.device_name = device_name
         self.imei=imei
+
+        # Uplink console
+        self.uplink_console = uplink_console
 
         #Flask server connection
         self.flask_server=tlm_config["webservice"]["server"]
@@ -88,9 +91,8 @@ class RadioSession(object):
             for i in range(len(fields)):
                 updated_fields[fields[i]]=vals[i]
 
-            #create a JSON file with the updated statefields and send it to the iridium email
-            with open('uplink.sbd', 'w') as json_uplink:
-                json.dump(updated_fields, json_uplink)
+            created_uplink = self.uplink_console.create_uplink(fields, vals, "uplink.sbd")
+            if not created_uplink: return False
             os.system("./ptest/send_uplink uplink.sbd")
             return True
         else:
